Use the new VisualizerRule from the corpus configuration to determine which visualizers to show.
This also changes the mapping type from "Properties" to the more general "Map<String, String>"

+++ annis-gui/src/main/java/annis/gui/resultview/VisualizerContextChanger.java
@@ -57,7 +57,7 @@ public interface VisualizerContextChanger extends Serializable {
      * @param status
      *            If true the visualizer is visible.
      */
-    public void registerVisibilityStatus(long entryId, boolean status);
+    public void registerVisibilityStatus(int entryId, boolean status);
 
     /**
      * Reinitiates all registered visualizer with a new salt project.
